Framework tested except 1.L10N, 2.DAO, 3.Memory tracking

+++ src/main/java/com/acl/qa/taf/util/ImageCompare.java
@@ -74,8 +74,8 @@ public class ImageCompare {
 		img2 = imageToBufferedImage(GrayFilter.createDisabledImage(img2));
 
 		// how big are each section
-		int blocksx = (int)(img1.getWidth() / comparex);
-		int blocksy = (int)(img1.getHeight() / comparey);
+		int blocksx = img1.getWidth() / comparex;
+		int blocksy = img1.getHeight() / comparey;
 
 		// set to a match by default, if a change is found then flag non-match
 		match = true;
@@ -115,7 +115,7 @@ public class ImageCompare {
 			}
 		}
 
-		return (int)(total / ((r.getWidth()/factorD)*(r.getHeight()/factorD)));
+		return total / ((r.getWidth()/factorD)*(r.getHeight()/factorD));
 	}
 
 	// constructor 1. use filenames
